feat(#77) : 프로필 subinfo 수정 반영
